server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Server API for the Virtual Shops eCommerce app"
authors = ["Sergio Barria"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
django-ninja = "^0.19.1"
python-dotenv = "^0.20.0"
Pillow = "^9.2.0"
pydantic = "^1.9.2"
psycopg2-binary = "^2.9.3"
dj-database-url = "^1.0.0"
django-storages = "^1.13.1"
boto3 = "^1.24.56"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
